glue-airflow/requirements.txt
apache-airflow-providers-amazon==1.3.0
boto3==1.17.54